+++ ChangeLog.md
@@ -1,3 +1,7 @@
+2019-09
+
+- prd.show/cb.show: allow for coefficient range selection and generate csv
+
 2018-09-27
 
 - complete first general revision over selected parts of